Changes Distances ArrayList type from Integer to Long. Closes #115

+++ server/src/main/java/com/tco/requests/Distances.java
@@ -1,13 +1,13 @@
 package com.tco.requests;
 import java.util.ArrayList;
 
-public class Distances extends ArrayList<Integer> {
+public class Distances extends ArrayList<Long> {
     
     public long total() {
         
         long totalDistance = 0;
 
-        for (Integer distance : this) {
+        for (Long distance : this) {
                 totalDistance += distance;
         }
